pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "requests>=2", "cibuildwheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsociety"
version = "2.0.0a98" # change it for each release
description = "LLM-based city environment agent building library"
authors = [
    { name = "Yuwei Yan", email = "pinkgranite86@gmail.com" },
    { name = "Junbo Yan", email = "yanjb20thu@gmali.com" },
    { name = "Jun Zhang", email = "zhangjun990222@gmali.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.10.10",
    "citystreetview>=1.2.4",
    "dashscope>=1.14.1",
    "geojson>=3.1.0",
    "grpcio>=1.67.1,<2.0.0",
    "matplotlib>=3.8.3",
    "networkx>=3.2.1",
    "numpy>=1.20.0,<2.0.0",
    "openai>=1.58.1",
    "Pillow>=11.0.0,<12.0.0",
    "protobuf<=4.24.0,<5.0.0",
    "pycitydata>=1.0.3",
    "pycityproto>=2.2.5",
    "requests>=2.32.3",
    "Shapely>=2.0.6",
    "PyYAML>=6.0.2",
    "zhipuai>=2.1.5.20230904",
    "mosstool>=1.3.0",
    "ray>=2.40.0",
    "aiomqtt>=2.3.0",
    "fastavro>=1.10.0",
    "pandavro>=1.8.0",
    "langchain-core>=0.3.28",
    "mlflow>=2.19.0",
    "psycopg[binary]>=3.2.3",
    "transformers>=4.47.1",
    "torch>=2.5.1",
    "faiss-cpu>=1.9.0.post1",
    "langchain-community>=0.3.13",
    "pydantic>=2.10.4",
]

[project.urls]
Homepage = "https://github.com/tsinghua-fib-lab/agentsociety"
Repository = "https://github.com/tsinghua-fib-lab/agentsociety.git"
Issues = "https://github.com/tsinghua-fib-lab/agentsociety/issues"

[project.scripts]
agentsociety-sim = "agentsociety.cli.wrapper:agentsociety_sim"
agentsociety-ui = "agentsociety.cli.wrapper:agentsociety_ui"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentsociety*"]

[tool.cibuildwheel]
build = ["cp3*manylinux_x86_64", "cp3*macosx_arm64"]
skip = ["cp36-*", "cp37-*", "cp38-*"]

[tool.cibuildwheel.linux]
repair-wheel-command = "auditwheel repair -w {dest_dir} {wheel}"

[tool.cibuildwheel.macos]
repair-wheel-command = "delocate-wheel -w {dest_dir} {wheel}"
